Code/requirements.txt
openai>=1.0.0
langgraph
fastmcp
requests
python-dotenv